+++ CHANGELOG.md
@@ -5,6 +5,9 @@ is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Changed
+- Create new directory for `tests`.
+
 ## [0.6.2](https://github.com/aescanes/scribe-of-lagash-visualization/releases/tag/0.6.2) - 2026-09-02
 
 ### Changed